Use 64-bit offsets in the pairwise SwiGLU backward kernel
At 275M MB16 on one rank the [tokens*top_k, 2*hidden] activation has
2.28e9 elements, so the int32 'base' offset overflowed and the first
backward produced non-finite gradients (grouped_mm arm crashed at step 1
with 'Non-finite loss encountered in OLMoDDPOptimizer'). The 8-GPU MB8
runs stayed under 2**31 and never hit it.

+++ src/olmo_core/ops/swiglu_pairwise.py
@@ -14,7 +14,11 @@ from triton.language.extra.cuda import libdevice
 
 @triton.jit
 def _swiglu_backward_pair(x, dy, dx, pairs, HIDDEN: tl.constexpr, BLOCK: tl.constexpr):
-    index = tl.program_id(0) * BLOCK + tl.arange(0, BLOCK)
+    # Offsets must be 64-bit: ``base`` reaches ``2 * pairs``, and a single rank's
+    # [tokens * top_k, 2 * hidden] activation exceeds 2**31 elements at large
+    # microbatches (e.g. 275M at MB16: 2,097,152 x 1088 = 2.28e9). With int32 the
+    # index wrapped negative and the first backward produced non-finite gradients.
+    index = tl.program_id(0).to(tl.int64) * BLOCK + tl.arange(0, BLOCK).to(tl.int64)
     mask = index < pairs
     base = index // HIDDEN * (2 * HIDDEN) + index % HIDDEN
     up = tl.load(x + base, mask=mask, other=0).to(tl.float32)
